Style experiment.
Note that this is certainly not final. I changed the fonts so some users
will see helvetica but others will see arial or even sans-serif. We can
change this but I think I definitely like it more now. I also removed
padding from elements on the homepage and added it to #content. The last
thing I tried was adding a shadow to the selected input boxes. I have
mixed feelings about it.