Host TeX images on our server.
Instead of using Google Charts API to serve our images for us we now use
our own server. The server must have latex and dvipng installed. The
images are created in a new thread when a user submits their post. This
works around a limitation in the length of input to Charts API. Images
are stored under assets/formulas/ using their sha1 has as a name.

Limitations:
  - The images are smaller than I would like
  - Formatting will likely be an issue
  - The threading needs some work. It probably isn't any better than
    without threads at this point.
  - There is a bug viewing previous revisions.