Host TeX formulas on our server
We now host TeX formulas on our server instead of using Google Charts
API. This works around a limitation in Charts API with the maximum
length of formulas. The Images are stored in assets/formulas with their
sha1 sum as a name. The server is required to have latex and dvipng
installed.

Limitations:
  - The formulas are smaller than I would like
  - Formatting may become an issue (as before)
  - Currently threading is implemented but probably provides no
    advantage
  - Their is a bug viewing old revisions (not new?)